added plot_downsampling code and refactored some utility stuff in the process
